+++ src/deepset_mcp/api/pipeline/__init__.py
@@ -1,5 +1,11 @@
-from .log_models import PipelineLog, PipelineLogList
-from .models import DeepsetPipeline, NoContentResponse, PipelineValidationResult, ValidationError
+from .models import (
+    DeepsetPipeline,
+    NoContentResponse,
+    PipelineLog,
+    PipelineLogList,
+    PipelineValidationResult,
+    ValidationError,
+)
 from .resource import PipelineResource
 
 __all__ = [
